Break up monolithic partial class
