+++ ICSharpCode.CodeConverter/VB/CommonConversions.cs
@@ -1,0 +1,450 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using ICSharpCode.CodeConverter.Util;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.VisualBasic;
+using Microsoft.CodeAnalysis.VisualBasic.Syntax;
+using AttributeListSyntax = Microsoft.CodeAnalysis.VisualBasic.Syntax.AttributeListSyntax;
+using CSharpExtensions = Microsoft.CodeAnalysis.CSharp.CSharpExtensions;
+using ExpressionSyntax = Microsoft.CodeAnalysis.VisualBasic.Syntax.ExpressionSyntax;
+using IdentifierNameSyntax = Microsoft.CodeAnalysis.VisualBasic.Syntax.IdentifierNameSyntax;
+using LambdaExpressionSyntax = Microsoft.CodeAnalysis.VisualBasic.Syntax.LambdaExpressionSyntax;
+using ParameterSyntax = Microsoft.CodeAnalysis.VisualBasic.Syntax.ParameterSyntax;
+using ReturnStatementSyntax = Microsoft.CodeAnalysis.VisualBasic.Syntax.ReturnStatementSyntax;
+using StatementSyntax = Microsoft.CodeAnalysis.VisualBasic.Syntax.StatementSyntax;
+using SyntaxFactory = Microsoft.CodeAnalysis.VisualBasic.SyntaxFactory;
+using SyntaxFacts = Microsoft.CodeAnalysis.VisualBasic.SyntaxFacts;
+using SyntaxKind = Microsoft.CodeAnalysis.VisualBasic.SyntaxKind;
+using TypeSyntax = Microsoft.CodeAnalysis.VisualBasic.Syntax.TypeSyntax;
+using VariableDeclaratorSyntax = Microsoft.CodeAnalysis.VisualBasic.Syntax.VariableDeclaratorSyntax;
+using YieldStatementSyntax = Microsoft.CodeAnalysis.VisualBasic.Syntax.YieldStatementSyntax;
+
+namespace ICSharpCode.CodeConverter.VB
+{
+    internal class CommonConversions
+    {
+        private readonly NodesVisitor _nodesVisitor;
+        private readonly SemanticModel _semanticModel;
+
+        public CommonConversions(SemanticModel semanticModel, NodesVisitor nodesVisitor)
+        {
+            _semanticModel = semanticModel;
+            _nodesVisitor = nodesVisitor;
+        }
+
+        public SyntaxList<StatementSyntax> ConvertBody(BlockSyntax body,
+            ArrowExpressionClauseSyntax expressionBody, MethodBodyVisitor iteratorState = null)
+        {
+            if (body != null) {
+                return ConvertStatements(body.Statements, iteratorState);
+            }
+
+            if (expressionBody != null) {
+                var convertedBody = expressionBody.Expression.Accept(_nodesVisitor.TriviaConvertingVisitor);
+                if (convertedBody is ExpressionSyntax convertedBodyExpression) {
+                    convertedBody = SyntaxFactory.ReturnStatement(convertedBodyExpression);
+                }
+
+                return SyntaxFactory.SingletonList((StatementSyntax)convertedBody);
+            }
+
+            return SyntaxFactory.List<StatementSyntax>();
+        }
+
+        private SyntaxList<StatementSyntax> ConvertStatements(SyntaxList<Microsoft.CodeAnalysis.CSharp.Syntax.StatementSyntax> statements,
+            MethodBodyVisitor iteratorState = null)
+        {
+            var methodBodyVisitor = CreateMethodBodyVisitor(iteratorState);
+            return SyntaxFactory.List(statements.SelectMany<Microsoft.CodeAnalysis.CSharp.Syntax.StatementSyntax, StatementSyntax>(s => ConvertStatement(s, methodBodyVisitor)));
+        }
+
+        private SyntaxList<StatementSyntax> ConvertStatement(Microsoft.CodeAnalysis.CSharp.Syntax.StatementSyntax statement, CSharpSyntaxVisitor<SyntaxList<StatementSyntax>> methodBodyVisitor)
+        {
+            var convertedStatements = statement.Accept(methodBodyVisitor);
+            convertedStatements = InsertRequiredDeclarations(convertedStatements, statement);
+
+            return convertedStatements;
+        }
+
+        private SyntaxList<StatementSyntax> InsertRequiredDeclarations(
+            SyntaxList<StatementSyntax> convertedStatements, CSharpSyntaxNode originaNode)
+        {
+            var declarationExpressions = originaNode.DescendantNodes().OfType<DeclarationExpressionSyntax>().ToList();
+            if (declarationExpressions.Any()) {
+                convertedStatements = convertedStatements.Insert(0, ConvertToDeclarationStatement(declarationExpressions));
+            }
+
+            return convertedStatements;
+        }
+
+        private StatementSyntax ConvertToDeclarationStatement(List<DeclarationExpressionSyntax> des)
+        {
+            var declarators = SyntaxFactory.SeparatedList(des.Select(ConvertToVariableDeclarator));
+            return SyntaxFactory.LocalDeclarationStatement(SyntaxFactory.TokenList(SyntaxFactory.Token(SyntaxKind.DimKeyword)), declarators);
+        }
+
+        private VariableDeclaratorSyntax ConvertToVariableDeclarator(DeclarationExpressionSyntax des)
+        {
+            var id = ((IdentifierNameSyntax)des.Accept(_nodesVisitor.TriviaConvertingVisitor)).Identifier;
+            var ids = SyntaxFactory.SingletonSeparatedList(SyntaxFactory.ModifiedIdentifier(id));
+            TypeSyntax typeSyntax;
+            if (des.Type.IsVar) {
+                var typeSymbol = (ITypeSymbol)ModelExtensions.GetSymbolInfo(_semanticModel, des.Type).ExtractBestMatch();
+                typeSyntax = typeSymbol.ToVbSyntax(_semanticModel, des.Type);
+            } else {
+                typeSyntax = (TypeSyntax)des.Type.Accept(_nodesVisitor.TriviaConvertingVisitor);
+            }
+
+            var simpleAsClauseSyntax = SyntaxFactory.SimpleAsClause(typeSyntax);
+            var equalsValueSyntax = SyntaxFactory.EqualsValue(SyntaxFactory.LiteralExpression(SyntaxKind.NothingLiteralExpression, SyntaxFactory.Token(SyntaxKind.NothingKeyword)));
+            return SyntaxFactory.VariableDeclarator(ids, simpleAsClauseSyntax, equalsValueSyntax);
+        }
+
+        private CSharpSyntaxVisitor<SyntaxList<StatementSyntax>> CreateMethodBodyVisitor(MethodBodyVisitor methodBodyVisitor = null)
+        {
+            var visitor = methodBodyVisitor ?? new MethodBodyVisitor(_semanticModel, _nodesVisitor.TriviaConvertingVisitor, _nodesVisitor.TriviaConvertingVisitor.TriviaConverter, this);
+            return visitor.CommentConvertingVisitor;
+        }
+
+        public AccessorBlockSyntax ConvertAccessor(AccessorDeclarationSyntax node, out bool isIterator)
+        {
+            SyntaxKind blockKind;
+            AccessorStatementSyntax stmt;
+            EndBlockStatementSyntax endStmt;
+            SyntaxList<StatementSyntax> body = SyntaxFactory.List<StatementSyntax>();
+            isIterator = false;
+            var isIteratorState = new MethodBodyVisitor(_semanticModel, _nodesVisitor.TriviaConvertingVisitor, _nodesVisitor.TriviaConvertingVisitor.TriviaConverter, this);
+            body = ConvertBody(node.Body, node.ExpressionBody, isIteratorState);
+            isIterator = isIteratorState.IsIterator;
+            var attributes = SyntaxFactory.List(node.AttributeLists.Select(a => (AttributeListSyntax)a.Accept(_nodesVisitor.TriviaConvertingVisitor)));
+            var modifiers = ConvertModifiers(node.Modifiers, SyntaxKindExtensions.TokenContext.Local);
+            var parent = (BasePropertyDeclarationSyntax)node.Parent.Parent;
+            ParameterSyntax valueParam;
+
+            switch (CSharpExtensions.Kind(node)) {
+                case Microsoft.CodeAnalysis.CSharp.SyntaxKind.GetAccessorDeclaration:
+                    blockKind = SyntaxKind.GetAccessorBlock;
+                    stmt = SyntaxFactory.GetAccessorStatement(attributes, modifiers, null);
+                    endStmt = SyntaxFactory.EndGetStatement();
+                    break;
+                case Microsoft.CodeAnalysis.CSharp.SyntaxKind.SetAccessorDeclaration:
+                    blockKind = SyntaxKind.SetAccessorBlock;
+                    valueParam = SyntaxFactory.Parameter(SyntaxFactory.ModifiedIdentifier("value"))
+                        .WithAsClause(SyntaxFactory.SimpleAsClause((TypeSyntax)parent.Type.Accept(_nodesVisitor.TriviaConvertingVisitor)))
+                        .WithModifiers(SyntaxFactory.TokenList(SyntaxFactory.Token(SyntaxKind.ByValKeyword)));
+                    stmt = SyntaxFactory.SetAccessorStatement(attributes, modifiers, SyntaxFactory.ParameterList(SyntaxFactory.SingletonSeparatedList(valueParam)));
+                    endStmt = SyntaxFactory.EndSetStatement();
+                    break;
+                case Microsoft.CodeAnalysis.CSharp.SyntaxKind.AddAccessorDeclaration:
+                    blockKind = SyntaxKind.AddHandlerAccessorBlock;
+                    valueParam = SyntaxFactory.Parameter(SyntaxFactory.ModifiedIdentifier("value"))
+                        .WithAsClause(SyntaxFactory.SimpleAsClause((TypeSyntax)parent.Type.Accept(_nodesVisitor.TriviaConvertingVisitor)))
+                        .WithModifiers(SyntaxFactory.TokenList(SyntaxFactory.Token(SyntaxKind.ByValKeyword)));
+                    stmt = SyntaxFactory.AddHandlerAccessorStatement(attributes, modifiers, SyntaxFactory.ParameterList(SyntaxFactory.SingletonSeparatedList(valueParam)));
+                    endStmt = SyntaxFactory.EndAddHandlerStatement();
+                    break;
+                case Microsoft.CodeAnalysis.CSharp.SyntaxKind.RemoveAccessorDeclaration:
+                    blockKind = SyntaxKind.RemoveHandlerAccessorBlock;
+                    valueParam = SyntaxFactory.Parameter(SyntaxFactory.ModifiedIdentifier("value"))
+                        .WithAsClause(SyntaxFactory.SimpleAsClause((TypeSyntax)parent.Type.Accept(_nodesVisitor.TriviaConvertingVisitor)))
+                        .WithModifiers(SyntaxFactory.TokenList(SyntaxFactory.Token(SyntaxKind.ByValKeyword)));
+                    stmt = SyntaxFactory.RemoveHandlerAccessorStatement(attributes, modifiers, SyntaxFactory.ParameterList(SyntaxFactory.SingletonSeparatedList(valueParam)));
+                    endStmt = SyntaxFactory.EndRemoveHandlerStatement();
+                    break;
+                default:
+                    throw new NotSupportedException();
+            }
+            return SyntaxFactory.AccessorBlock(blockKind, stmt, body, endStmt);
+        }
+
+        public ExpressionSyntax ReduceArrayUpperBoundExpression(Microsoft.CodeAnalysis.CSharp.Syntax.ExpressionSyntax expr)
+        {
+            var constant = _semanticModel.GetConstantValue(expr);
+            if (constant.HasValue && constant.Value is int)
+                return SyntaxFactory.NumericLiteralExpression(SyntaxFactory.Literal((int)constant.Value - 1));
+
+            return SyntaxFactory.BinaryExpression(
+                SyntaxKind.SubtractExpression,
+                (ExpressionSyntax)expr.Accept(_nodesVisitor.TriviaConvertingVisitor), SyntaxFactory.Token(SyntaxKind.MinusToken), SyntaxFactory.NumericLiteralExpression(SyntaxFactory.Literal(1)));
+        }
+
+        public LambdaExpressionSyntax ConvertLambdaExpression(AnonymousFunctionExpressionSyntax node, CSharpSyntaxNode body, IEnumerable<Microsoft.CodeAnalysis.CSharp.Syntax.ParameterSyntax> parameters, SyntaxTokenList modifiers)
+        {
+            var symbol = ModelExtensions.GetSymbolInfo(_semanticModel, node).Symbol as IMethodSymbol;
+            var parameterList = SyntaxFactory.ParameterList(SyntaxFactory.SeparatedList(parameters.Select(p => (ParameterSyntax)p.Accept(_nodesVisitor.TriviaConvertingVisitor))));
+            LambdaHeaderSyntax header;
+            EndBlockStatementSyntax endBlock;
+            SyntaxKind multiLineExpressionKind;
+            SyntaxKind singleLineExpressionKind;
+            if (symbol.ReturnsVoid) {
+                header = SyntaxFactory.SubLambdaHeader(SyntaxFactory.List<AttributeListSyntax>(),
+                    ConvertModifiers(modifiers, SyntaxKindExtensions.TokenContext.Local), parameterList, null);
+                endBlock = SyntaxFactory.EndBlockStatement(SyntaxKind.EndSubStatement,
+                    SyntaxFactory.Token(SyntaxKind.SubKeyword));
+                multiLineExpressionKind = SyntaxKind.MultiLineSubLambdaExpression;
+                singleLineExpressionKind = SyntaxKind.SingleLineSubLambdaExpression;
+            } else {
+                header = SyntaxFactory.FunctionLambdaHeader(SyntaxFactory.List<AttributeListSyntax>(),
+                    ConvertModifiers(modifiers, SyntaxKindExtensions.TokenContext.Local), parameterList, null);
+                endBlock = SyntaxFactory.EndBlockStatement(SyntaxKind.EndFunctionStatement, SyntaxFactory.Token(SyntaxKind.FunctionKeyword));
+                multiLineExpressionKind = SyntaxKind.MultiLineFunctionLambdaExpression;
+                singleLineExpressionKind = SyntaxKind.SingleLineFunctionLambdaExpression;
+            }
+
+            SyntaxList<StatementSyntax> statements;
+            if (body is BlockSyntax block) {
+                statements = ConvertStatements(block.Statements);
+            } else {
+                statements = InsertRequiredDeclarations(
+                    SyntaxFactory.SingletonList<StatementSyntax>(
+                        SyntaxFactory.ReturnStatement((ExpressionSyntax)body.Accept(_nodesVisitor.TriviaConvertingVisitor))),
+                    body);
+            }
+
+            if (statements.Count == 1 && UnpackExpressionFromStatement(statements[0], out var expression)) {
+                return SyntaxFactory.SingleLineLambdaExpression(singleLineExpressionKind, header, expression);
+            }
+
+            return SyntaxFactory.MultiLineLambdaExpression(multiLineExpressionKind, header, statements, endBlock);
+        }
+
+        private bool UnpackExpressionFromStatement(StatementSyntax statementSyntax, out ExpressionSyntax expression)
+        {
+            if (statementSyntax is ReturnStatementSyntax)
+                expression = ((ReturnStatementSyntax)statementSyntax).Expression;
+            else if (statementSyntax is YieldStatementSyntax)
+                expression = ((YieldStatementSyntax)statementSyntax).Expression;
+            else
+                expression = null;
+            return expression != null;
+        }
+
+        public void ConvertBaseList(BaseTypeDeclarationSyntax type, List<InheritsStatementSyntax> inherits, List<ImplementsStatementSyntax> implements)
+        {
+            TypeSyntax[] arr;
+            switch (type.Kind()) {
+                case Microsoft.CodeAnalysis.CSharp.SyntaxKind.ClassDeclaration:
+                    var classOrInterface = type.BaseList?.Types.FirstOrDefault()?.Type;
+                    if (classOrInterface == null) return;
+                    var classOrInterfaceSymbol = ModelExtensions.GetSymbolInfo(_semanticModel, classOrInterface).Symbol;
+                    if (classOrInterfaceSymbol?.IsInterfaceType() == true) {
+                        arr = type.BaseList?.Types.Select(t => (TypeSyntax)t.Type.Accept(_nodesVisitor.TriviaConvertingVisitor)).ToArray();
+                        if (arr.Length > 0)
+                            implements.Add(SyntaxFactory.ImplementsStatement(arr));
+                    } else {
+                        inherits.Add(SyntaxFactory.InheritsStatement((TypeSyntax)classOrInterface.Accept(_nodesVisitor.TriviaConvertingVisitor)));
+                        arr = type.BaseList?.Types.Skip(1).Select(t => (TypeSyntax)t.Type.Accept(_nodesVisitor.TriviaConvertingVisitor)).ToArray();
+                        if (arr.Length > 0)
+                            implements.Add(SyntaxFactory.ImplementsStatement(arr));
+                    }
+                    break;
+                case Microsoft.CodeAnalysis.CSharp.SyntaxKind.StructDeclaration:
+                    arr = type.BaseList?.Types.Select(t => (TypeSyntax)t.Type.Accept(_nodesVisitor.TriviaConvertingVisitor)).ToArray();
+                    if (arr?.Length > 0)
+                        implements.Add(SyntaxFactory.ImplementsStatement(arr));
+                    break;
+                case Microsoft.CodeAnalysis.CSharp.SyntaxKind.InterfaceDeclaration:
+                    arr = type.BaseList?.Types.Select(t => (TypeSyntax)t.Type.Accept(_nodesVisitor.TriviaConvertingVisitor)).ToArray();
+                    if (arr?.Length > 0)
+                        inherits.Add(SyntaxFactory.InheritsStatement(arr));
+                    break;
+            }
+        }
+
+
+        private static IEnumerable<SyntaxToken> ConvertModifiersCore(IEnumerable<SyntaxToken> modifiers, SyntaxKindExtensions.TokenContext context)
+        {
+            if (context != SyntaxKindExtensions.TokenContext.Local && context != SyntaxKindExtensions.TokenContext.InterfaceOrModule) {
+                bool visibility = false;
+                foreach (var token in modifiers) {
+                    if (IsVisibility(token, context)) {
+                        visibility = true;
+                        break;
+                    }
+                }
+                if (!visibility && context == SyntaxKindExtensions.TokenContext.Member)
+                    yield return CSharpDefaultVisibility(context);
+            }
+            foreach (var token in modifiers.Where(m => !IgnoreInContext(m, context))) {
+                var m = ConvertModifier(token, context);
+                if (m.HasValue) yield return m.Value;
+            }
+        }
+
+        private static bool IgnoreInContext(SyntaxToken m, SyntaxKindExtensions.TokenContext context)
+        {
+            switch (context) {
+                case SyntaxKindExtensions.TokenContext.InterfaceOrModule:
+                    return m.IsKind(Microsoft.CodeAnalysis.CSharp.SyntaxKind.PublicKeyword, Microsoft.CodeAnalysis.CSharp.SyntaxKind.StaticKeyword);
+            }
+            return false;
+        }
+
+        private static bool IsVisibility(SyntaxToken token, SyntaxKindExtensions.TokenContext context)
+        {
+            return token.IsKind(Microsoft.CodeAnalysis.CSharp.SyntaxKind.PublicKeyword, Microsoft.CodeAnalysis.CSharp.SyntaxKind.InternalKeyword, Microsoft.CodeAnalysis.CSharp.SyntaxKind.ProtectedKeyword, Microsoft.CodeAnalysis.CSharp.SyntaxKind.PrivateKeyword)
+                   || (context == SyntaxKindExtensions.TokenContext.VariableOrConst && SyntaxTokenExtensions.IsKind(token, Microsoft.CodeAnalysis.CSharp.SyntaxKind.ConstKeyword));
+        }
+
+        private static SyntaxToken CSharpDefaultVisibility(SyntaxKindExtensions.TokenContext context)
+        {
+            switch (context) {
+                case SyntaxKindExtensions.TokenContext.Global:
+                    return SyntaxFactory.Token(SyntaxKind.FriendKeyword);
+                case SyntaxKindExtensions.TokenContext.Local:
+                case SyntaxKindExtensions.TokenContext.VariableOrConst:
+                case SyntaxKindExtensions.TokenContext.Member:
+                    return SyntaxFactory.Token(SyntaxKind.PrivateKeyword);
+            }
+            throw new ArgumentOutOfRangeException(nameof(context));
+        }
+
+        internal static SyntaxTokenList ConvertModifiers(IEnumerable<SyntaxToken> modifiers, SyntaxKindExtensions.TokenContext context = SyntaxKindExtensions.TokenContext.Global)
+        {
+            return SyntaxFactory.TokenList(ConvertModifiersCore(modifiers, context));
+        }
+
+        private static SyntaxToken? ConvertModifier(SyntaxToken m, SyntaxKindExtensions.TokenContext context = SyntaxKindExtensions.TokenContext.Global)
+        {
+            var token = SyntaxKindExtensions.ConvertToken(Microsoft.CodeAnalysis.CSharp.CSharpExtensions.Kind(m), context);
+            return token == SyntaxKind.None ? null : new SyntaxToken?(SyntaxFactory.Token(token));
+        }
+
+        internal SeparatedSyntaxList<VariableDeclaratorSyntax> RemodelVariableDeclaration(Microsoft.CodeAnalysis.CSharp.Syntax.VariableDeclarationSyntax declaration, Microsoft.CodeAnalysis.CSharp.CSharpSyntaxVisitor<VisualBasicSyntaxNode> nodesVisitor)
+        {
+            var visualBasicSyntaxNode = declaration.Type.Accept(nodesVisitor);
+            var type = (TypeSyntax)visualBasicSyntaxNode;
+            var declaratorsWithoutInitializers = new List<Microsoft.CodeAnalysis.CSharp.Syntax.VariableDeclaratorSyntax>();
+            var declarators = new List<VariableDeclaratorSyntax>();
+
+            foreach (var v in declaration.Variables) {
+                if (v.Initializer == null) {
+                    declaratorsWithoutInitializers.Add(v);
+                    continue;
+                } else {
+                    declarators.Add(
+                        SyntaxFactory.VariableDeclarator(
+                            SyntaxFactory.SingletonSeparatedList<ModifiedIdentifierSyntax>(ExtractIdentifier(v)),
+                            declaration.Type.IsVar ? null : SyntaxFactory.SimpleAsClause(type),
+                            SyntaxFactory.EqualsValue((ExpressionSyntax)ConvertTopLevelExpression(v.Initializer.Value, nodesVisitor))
+                        )
+                    );
+                }
+            }
+
+            if (declaratorsWithoutInitializers.Count > 0) {
+                declarators.Insert(0, SyntaxFactory.VariableDeclarator(SyntaxFactory.SeparatedList(declaratorsWithoutInitializers.Select(ExtractIdentifier)), SyntaxFactory.SimpleAsClause(type), null));
+            }
+
+            return SyntaxFactory.SeparatedList(declarators);
+        }
+
+        private static VisualBasicSyntaxNode ConvertTopLevelExpression(Microsoft.CodeAnalysis.CSharp.Syntax.ExpressionSyntax topLevelExpression,
+            Microsoft.CodeAnalysis.CSharp.CSharpSyntaxVisitor<VisualBasicSyntaxNode> nodesVisitor)
+        {
+            return topLevelExpression.Accept(nodesVisitor);
+        }
+
+        private static ModifiedIdentifierSyntax ExtractIdentifier(Microsoft.CodeAnalysis.CSharp.Syntax.VariableDeclaratorSyntax v)
+        {
+            return SyntaxFactory.ModifiedIdentifier((SyntaxToken)ConvertIdentifier(v.Identifier));
+        }
+
+        public static SyntaxToken ConvertIdentifier(SyntaxToken id)
+        {
+            var idText = id.ValueText;
+            // Underscore is a special character in VB lexer which continues lines - not sure where to find the whole set of other similar tokens if any
+            // Rather than a complicated contextual rename, just add an extra dash to all identifiers and hope this method is consistently used
+            if (idText.All(c => c == '_')) idText += "_";
+            var keywordKind = SyntaxFacts.GetKeywordKind(idText);
+            if (keywordKind != SyntaxKind.None && !SyntaxFacts.IsPredefinedType(keywordKind))
+                return SyntaxFactory.Identifier("[" + idText + "]");
+            return SyntaxFactory.Identifier(idText);
+        }
+
+        public static ExpressionSyntax Literal(object o, string valueText = null) => GetLiteralExpression(o, valueText);
+
+        internal static ExpressionSyntax GetLiteralExpression(object value, string valueText = null)
+        {
+            if (value is char)
+                return SyntaxFactory.LiteralExpression(SyntaxKind.CharacterLiteralExpression, SyntaxFactory.Literal((char)value));
+
+            if (value is string)
+                return SyntaxFactory.LiteralExpression(SyntaxKind.StringLiteralExpression, SyntaxFactory.Literal((string)value));
+
+            if (value == null)
+                return SyntaxFactory.NothingLiteralExpression(SyntaxFactory.Token(SyntaxKind.NothingKeyword));
+
+            if (value is bool)
+                return (bool)value ? SyntaxFactory.TrueLiteralExpression(SyntaxFactory.Token(SyntaxKind.TrueKeyword)) : SyntaxFactory.FalseLiteralExpression(SyntaxFactory.Token(SyntaxKind.FalseKeyword));
+
+
+            valueText = valueText != null ? ConvertNumericLiteralValueText(valueText) : value.ToString();
+
+            if (value is byte)
+                return SyntaxFactory.LiteralExpression(SyntaxKind.NumericLiteralExpression, SyntaxFactory.Literal(valueText, (byte)value));
+            if (value is sbyte)
+                return SyntaxFactory.LiteralExpression(SyntaxKind.NumericLiteralExpression, SyntaxFactory.Literal(valueText, (sbyte)value));
+            if (value is short)
+                return SyntaxFactory.LiteralExpression(SyntaxKind.NumericLiteralExpression, SyntaxFactory.Literal(valueText, (short)value));
+            if (value is ushort)
+                return SyntaxFactory.LiteralExpression(SyntaxKind.NumericLiteralExpression, SyntaxFactory.Literal(valueText, (ushort)value));
+            if (value is int)
+                return SyntaxFactory.LiteralExpression(SyntaxKind.NumericLiteralExpression, SyntaxFactory.Literal(valueText, (int)value));
+            if (value is uint)
+                return SyntaxFactory.LiteralExpression(SyntaxKind.NumericLiteralExpression, SyntaxFactory.Literal(valueText, (uint)value));
+            if (value is long)
+                return SyntaxFactory.LiteralExpression(SyntaxKind.NumericLiteralExpression, SyntaxFactory.Literal(valueText, (long)value));
+            if (value is ulong)
+                return SyntaxFactory.LiteralExpression(SyntaxKind.NumericLiteralExpression, SyntaxFactory.Literal(valueText, (ulong)value));
+
+            if (value is float)
+                return SyntaxFactory.LiteralExpression(SyntaxKind.NumericLiteralExpression, SyntaxFactory.Literal(valueText, (float)value));
+            if (value is double)
+                return SyntaxFactory.LiteralExpression(SyntaxKind.NumericLiteralExpression, SyntaxFactory.Literal(valueText, (double)value));
+            if (value is decimal)
+                return SyntaxFactory.LiteralExpression(SyntaxKind.NumericLiteralExpression, SyntaxFactory.Literal(valueText, (decimal)value));
+
+            throw new ArgumentOutOfRangeException(nameof(value), value, null);
+        }
+
+
+        /// <summary>
+        ///  https://docs.microsoft.com/en-us/dotnet/visual-basic/programming-guide/language-features/data-types/type-characters
+        /// </summary>
+        private static string ConvertNumericLiteralValueText(string valueText)
+        {
+            var replacements = new Dictionary<string, string> {
+                {"U", "UI"},
+                {"UL", "UL"},
+                {"M", "D"},
+                {"F", "F"},
+                {"D", "R"},
+                {"L", "L"}, // Normalizes casing
+            };
+
+            // Be careful not to replace only the "L" in "UL" for example
+            var longestMatchingReplacement = replacements.Where(t => valueText.EndsWith(t.Key, StringComparison.OrdinalIgnoreCase))
+                .GroupBy(t => t.Key.Length).OrderByDescending(g => g.Key).FirstOrDefault()?.SingleOrDefault();
+
+            if (longestMatchingReplacement != null) {
+                valueText = valueText.ReplaceEnd(longestMatchingReplacement.Value);
+            }
+
+            if (valueText.Length <= 2) return valueText;
+
+            if (valueText.StartsWith("0x")) {
+                return "&H" + valueText.Substring(2).Replace("R", "D"); // Undo any accidental replacements that assumed this was a decimal;
+            }
+
+            if (valueText.StartsWith("0b")) {
+                return "&B" + valueText.Substring(2);
+            }
+
+            return valueText;
+        }
+    }
+}
